rename package 'shared' to 'models'
