+++ pkg/models/query.go
@@ -1,0 +1,6 @@
+package models
+
+type QueryModel struct {
+	QueryType string `json:"queryType"`
+	RawSql    string `json:"rawSql"`
+}
