+++ pkg/models/settings.go
@@ -1,0 +1,8 @@
+package models
+
+type Settings struct {
+	Url      string
+	User     string
+	Database string
+	Password string
+}
